requirements.txt
streamlit
streamlit-authenticator
st-pages

transformers
librosa

farm-haystack[ocr,preprocessing,file-conversion,pdf]
sentencepiece
sacremoses
